Reorg zenoh lib, fix ci, start on service
