+++ roslibrust/examples/generic_message.rs
@@ -7,14 +7,12 @@ use roslibrust_codegen::RosMessageType;
 /// We place the ros1 generate code into a module to prevent name collisions with the identically
 /// named ros2 types.
 mod ros1 {
-    roslibrust_codegen_macro::find_and_generate_ros_messages!(
-        "assets/ros1_common_interfaces/std_msgs"
-    );
+    roslibrust_codegen_macro::find_and_generate_ros_messages!("assets/ros1_common_interfaces");
 }
 
 mod ros2 {
     roslibrust_codegen_macro::find_and_generate_ros_messages_without_ros_package_path!(
-        "assets/ros2_common_interfaces/std_msgs"
+        "assets/ros2_common_interfaces"
     );
 }
